Add the possibility to suppress diagnostics based on abs path of the file
the paths are matched with a std::regex.
This is targetting this usecase.
Imagine to have a toolkit and you do not want to run a particular checker
on all the files of a certain directory in its source tree. You can express
this in a regex, specified in the checker traits.
If the regex is not specified, no file name is extracted and no matching is carried
out for performance reasons.